refactor(export): split launcher_store into per-domain submodules
Split the 1565-line launcher_store.rs into a directory module with per-
domain submodules (types, paths, fs_ops, queries, mutations) plus a
tests/ subdirectory with per-operation test files. Public re-export
surface in export/mod.rs is preserved byte-identical; all existing
tests pass unchanged.

Part of #290 (Tier 1, task 10).
