--- src/crosshook-native/crates/crosshook-core/src/export/launcher_store.rs
@@ -1,1565 +1,0 @@
-use std::error::Error;
-use std::fmt;
-use std::fs;
-use std::io;
-use std::path::Path;
-
-use serde::{Deserialize, Serialize};
-
-use super::launcher::{
-    build_desktop_entry_content, build_trainer_script_content, combine_host_unix_path,
-    resolve_display_name, resolve_target_home_path, sanitize_launcher_slug, strip_trainer_suffix,
-    write_host_text_file, SteamExternalLauncherExportRequest,
-};
-use crate::profile::{resolve_launch_method, GameProfile};
-use crate::settings::UmuPreference;
-
-/// Watermark present in generated `.sh` trainer launcher scripts.
-const SCRIPT_WATERMARK: &str = "# Generated by CrossHook";
-
-/// Watermark present in generated `.desktop` entry files.
-const DESKTOP_ENTRY_WATERMARK: &str = "Generated by CrossHook";
-
-/// Metadata about an exported launcher pair on disk.
-///
-/// `is_stale` is only meaningful when the value was derived with profile context
-/// via `check_launcher_exists` / `check_launcher_for_profile`. `list_launchers`
-/// does not have that context and currently reports `is_stale = false`.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub struct LauncherInfo {
-    #[serde(default)]
-    pub display_name: String,
-    #[serde(default)]
-    pub launcher_slug: String,
-    #[serde(default)]
-    pub script_path: String,
-    #[serde(default)]
-    pub desktop_entry_path: String,
-    #[serde(default)]
-    pub script_exists: bool,
-    #[serde(default)]
-    pub desktop_entry_exists: bool,
-    #[serde(default)]
-    pub is_stale: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub struct LauncherDeleteResult {
-    #[serde(default)]
-    pub script_deleted: bool,
-    #[serde(default)]
-    pub desktop_entry_deleted: bool,
-    #[serde(default)]
-    pub script_path: String,
-    #[serde(default)]
-    pub desktop_entry_path: String,
-    #[serde(default)]
-    pub script_skipped_reason: Option<String>,
-    #[serde(default)]
-    pub desktop_entry_skipped_reason: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub struct LauncherRenameResult {
-    #[serde(default)]
-    pub old_slug: String,
-    #[serde(default)]
-    pub new_slug: String,
-    #[serde(default)]
-    pub new_script_path: String,
-    #[serde(default)]
-    pub new_desktop_entry_path: String,
-    #[serde(default)]
-    pub script_renamed: bool,
-    #[serde(default)]
-    pub desktop_entry_renamed: bool,
-    #[serde(default)]
-    pub old_script_cleanup_warning: Option<String>,
-    #[serde(default)]
-    pub old_desktop_entry_cleanup_warning: Option<String>,
-}
-
-#[derive(Debug)]
-pub enum LauncherStoreError {
-    Io(io::Error),
-    HomePathResolutionFailed,
-}
-
-impl fmt::Display for LauncherStoreError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Io(error) => write!(f, "{error}"),
-            Self::HomePathResolutionFailed => {
-                f.write_str("Could not resolve a host home path for launcher operations.")
-            }
-        }
-    }
-}
-
-impl Error for LauncherStoreError {
-    fn source(&self) -> Option<&(dyn Error + 'static)> {
-        match self {
-            Self::Io(error) => Some(error),
-            _ => None,
-        }
-    }
-}
-
-impl From<io::Error> for LauncherStoreError {
-    fn from(value: io::Error) -> Self {
-        Self::Io(value)
-    }
-}
-
-/// Derives the resolved display name, launcher slug, script path, and desktop entry path
-/// from the given inputs. Shared by `check_launcher_exists` and `delete_launcher_files`.
-fn derive_launcher_paths(
-    display_name: &str,
-    steam_app_id: &str,
-    trainer_path: &str,
-    target_home_path: &str,
-    steam_client_install_path: &str,
-) -> (String, String, String, String) {
-    let resolved_name = resolve_display_name(display_name, steam_app_id, trainer_path);
-    let slug = sanitize_launcher_slug(&resolved_name);
-    let home = resolve_target_home_path(target_home_path, steam_client_install_path);
-
-    let script_path = combine_host_unix_path(
-        &home,
-        ".local/share/crosshook/launchers",
-        &format!("{slug}-trainer.sh"),
-    );
-    let desktop_entry_path = combine_host_unix_path(
-        &home,
-        ".local/share/applications",
-        &format!("crosshook-{slug}-trainer.desktop"),
-    );
-
-    (resolved_name, slug, script_path, desktop_entry_path)
-}
-
-fn derive_launcher_paths_from_slug(
-    launcher_slug: &str,
-    target_home_path: &str,
-    steam_client_install_path: &str,
-) -> (String, String) {
-    let home = resolve_target_home_path(target_home_path, steam_client_install_path);
-    let normalized_slug = sanitize_launcher_slug(launcher_slug);
-
-    let script_path = combine_host_unix_path(
-        &home,
-        ".local/share/crosshook/launchers",
-        &format!("{normalized_slug}-trainer.sh"),
-    );
-    let desktop_entry_path = combine_host_unix_path(
-        &home,
-        ".local/share/applications",
-        &format!("crosshook-{normalized_slug}-trainer.desktop"),
-    );
-
-    (script_path, desktop_entry_path)
-}
-
-fn delete_launcher_at_paths(
-    script_path: String,
-    desktop_entry_path: String,
-) -> Result<LauncherDeleteResult, LauncherStoreError> {
-    let mut result = LauncherDeleteResult {
-        script_path: script_path.clone(),
-        desktop_entry_path: desktop_entry_path.clone(),
-        ..Default::default()
-    };
-
-    // Delete desktop entry first (user-visible artifact), with watermark verification
-    match verify_crosshook_file(&desktop_entry_path, DESKTOP_ENTRY_WATERMARK)? {
-        Some(reason) => {
-            result.desktop_entry_skipped_reason = Some(reason);
-        }
-        None => {
-            result.desktop_entry_deleted = remove_file_if_exists(&desktop_entry_path)?;
-        }
-    }
-
-    // Delete script, with watermark verification
-    match verify_crosshook_file(&script_path, SCRIPT_WATERMARK)? {
-        Some(reason) => {
-            result.script_skipped_reason = Some(reason);
-        }
-        None => {
-            result.script_deleted = remove_file_if_exists(&script_path)?;
-        }
-    }
-
-    Ok(result)
-}
-
-pub fn check_launcher_exists(
-    display_name: &str,
-    steam_app_id: &str,
-    trainer_path: &str,
-    target_home_path: &str,
-    steam_client_install_path: &str,
-) -> Result<LauncherInfo, LauncherStoreError> {
-    let (resolved_name, slug, script_path, desktop_entry_path) = derive_launcher_paths(
-        display_name,
-        steam_app_id,
-        trainer_path,
-        target_home_path,
-        steam_client_install_path,
-    );
-
-    let script_exists = !script_path.is_empty() && Path::new(&script_path).exists();
-    let desktop_entry_exists =
-        !desktop_entry_path.is_empty() && Path::new(&desktop_entry_path).exists();
-
-    // Check staleness: compare Name= line in .desktop against expected
-    let is_stale = if desktop_entry_exists {
-        match extract_display_name_from_desktop(&desktop_entry_path)? {
-            Some(actual_name) => actual_name != resolved_name,
-            None => true,
-        }
-    } else {
-        false
-    };
-
-    Ok(LauncherInfo {
-        display_name: resolved_name,
-        launcher_slug: slug,
-        script_path,
-        desktop_entry_path,
-        script_exists,
-        desktop_entry_exists,
-        is_stale,
-    })
-}
-
-pub fn check_launcher_exists_for_request(
-    display_name: &str,
-    request: &SteamExternalLauncherExportRequest,
-) -> Result<LauncherInfo, LauncherStoreError> {
-    let (resolved_name, slug, script_path, desktop_entry_path) = derive_launcher_paths(
-        display_name,
-        &request.steam_app_id,
-        &request.trainer_path,
-        &request.target_home_path,
-        &request.steam_client_install_path,
-    );
-
-    let script_exists = !script_path.is_empty() && Path::new(&script_path).exists();
-    let desktop_entry_exists =
-        !desktop_entry_path.is_empty() && Path::new(&desktop_entry_path).exists();
-
-    let script_is_stale = if script_exists {
-        let expected_script = build_trainer_script_content(request, &resolved_name);
-        fs::read_to_string(&script_path)
-            .map(|actual_script| actual_script != expected_script)
-            .unwrap_or(true)
-    } else {
-        false
-    };
-
-    let desktop_is_stale = if desktop_entry_exists {
-        match extract_display_name_from_desktop(&desktop_entry_path)? {
-            Some(actual_name) => actual_name != resolved_name,
-            None => true,
-        }
-    } else {
-        false
-    };
-
-    Ok(LauncherInfo {
-        display_name: resolved_name,
-        launcher_slug: slug,
-        script_path,
-        desktop_entry_path,
-        script_exists,
-        desktop_entry_exists,
-        is_stale: script_is_stale || desktop_is_stale,
-    })
-}
-
-pub fn check_launcher_for_profile(
-    profile: &GameProfile,
-    target_home_path: &str,
-    steam_client_install_path: &str,
-    global_umu_preference: UmuPreference,
-) -> Result<LauncherInfo, LauncherStoreError> {
-    let resolved_method = resolve_launch_method(profile);
-
-    if resolved_method == "native" {
-        return Ok(LauncherInfo::default());
-    }
-
-    check_launcher_exists_for_request(
-        &profile.steam.launcher.display_name,
-        &SteamExternalLauncherExportRequest {
-            method: resolved_method.to_string(),
-            launcher_name: profile.steam.launcher.display_name.clone(),
-            trainer_path: profile.trainer.path.clone(),
-            trainer_loading_mode: profile.trainer.loading_mode,
-            launcher_icon_path: profile.steam.launcher.icon_path.clone(),
-            prefix_path: if resolved_method == "steam_applaunch" {
-                profile.steam.compatdata_path.clone()
-            } else {
-                profile.runtime.prefix_path.clone()
-            },
-            proton_path: if resolved_method == "steam_applaunch" {
-                profile.steam.proton_path.clone()
-            } else {
-                profile.runtime.proton_path.clone()
-            },
-            steam_app_id: profile.steam.app_id.clone(),
-            steam_client_install_path: steam_client_install_path.to_string(),
-            target_home_path: target_home_path.to_string(),
-            profile_name: None,
-            runtime_steam_app_id: profile.runtime.steam_app_id.clone(),
-            umu_game_id: profile.runtime.umu_game_id.clone(),
-            umu_preference: profile
-                .runtime
-                .umu_preference
-                .unwrap_or(global_umu_preference),
-            network_isolation: profile.launch.network_isolation,
-            gamescope: profile.launch.resolved_trainer_gamescope(),
-        },
-    )
-}
-
-pub fn delete_launcher_files(
-    display_name: &str,
-    steam_app_id: &str,
-    trainer_path: &str,
-    target_home_path: &str,
-    steam_client_install_path: &str,
-) -> Result<LauncherDeleteResult, LauncherStoreError> {
-    let (_resolved_name, _slug, script_path, desktop_entry_path) = derive_launcher_paths(
-        display_name,
-        steam_app_id,
-        trainer_path,
-        target_home_path,
-        steam_client_install_path,
-    );
-
-    delete_launcher_at_paths(script_path, desktop_entry_path)
-}
-
-pub fn delete_launcher_by_slug(
-    launcher_slug: &str,
-    target_home_path: &str,
-    steam_client_install_path: &str,
-) -> Result<LauncherDeleteResult, LauncherStoreError> {
-    let (script_path, desktop_entry_path) =
-        derive_launcher_paths_from_slug(launcher_slug, target_home_path, steam_client_install_path);
-    delete_launcher_at_paths(script_path, desktop_entry_path)
-}
-
-pub fn delete_launcher_for_profile(
-    profile: &GameProfile,
-    target_home_path: &str,
-    steam_client_install_path: &str,
-) -> Result<LauncherDeleteResult, LauncherStoreError> {
-    let display_name = &profile.steam.launcher.display_name;
-    let steam_app_id = &profile.steam.app_id;
-    let trainer_path = &profile.trainer.path;
-
-    delete_launcher_files(
-        display_name,
-        steam_app_id,
-        trainer_path,
-        target_home_path,
-        steam_client_install_path,
-    )
-}
-
-/// Renames launcher files from one slug to another, rewriting file content with updated
-/// display names and paths. Uses a write-then-delete strategy because both `.sh` and
-/// `.desktop` files embed display names and paths as plaintext.
-///
-/// When the slug is unchanged, files are rewritten in place and the old paths are not deleted.
-pub fn rename_launcher_files(
-    old_launcher_slug: &str,
-    new_display_name: &str,
-    new_launcher_icon_path: &str,
-    target_home_path: &str,
-    steam_client_install_path: &str,
-    request: &SteamExternalLauncherExportRequest,
-) -> Result<LauncherRenameResult, LauncherStoreError> {
-    let new_slug = sanitize_launcher_slug(new_display_name);
-    let home = resolve_target_home_path(target_home_path, steam_client_install_path);
-
-    // Construct old file paths from old_launcher_slug
-    let old_script_path = combine_host_unix_path(
-        &home,
-        ".local/share/crosshook/launchers",
-        &format!("{old_launcher_slug}-trainer.sh"),
-    );
-    let old_desktop_entry_path = combine_host_unix_path(
-        &home,
-        ".local/share/applications",
-        &format!("crosshook-{old_launcher_slug}-trainer.desktop"),
-    );
-
-    // Construct new file paths from new_slug
-    let new_script_path = combine_host_unix_path(
-        &home,
-        ".local/share/crosshook/launchers",
-        &format!("{new_slug}-trainer.sh"),
-    );
-    let new_desktop_entry_path = combine_host_unix_path(
-        &home,
-        ".local/share/applications",
-        &format!("crosshook-{new_slug}-trainer.desktop"),
-    );
-
-    // Check if old files exist. If neither exists, return early with renamed: false
-    let old_script_exists = !old_script_path.is_empty() && Path::new(&old_script_path).exists();
-    let old_desktop_exists =
-        !old_desktop_entry_path.is_empty() && Path::new(&old_desktop_entry_path).exists();
-
-    if !old_script_exists && !old_desktop_exists {
-        return Ok(LauncherRenameResult {
-            old_slug: old_launcher_slug.to_string(),
-            new_slug,
-            new_script_path,
-            new_desktop_entry_path,
-            script_renamed: false,
-            desktop_entry_renamed: false,
-            old_script_cleanup_warning: None,
-            old_desktop_entry_cleanup_warning: None,
-        });
-    }
-
-    // Generate new file content
-    let new_script_content = build_trainer_script_content(request, new_display_name);
-    let new_desktop_content = build_desktop_entry_content(
-        new_display_name,
-        &new_slug,
-        &new_script_path,
-        new_launcher_icon_path,
-    );
-
-    // Write new files
-    let mut script_renamed = false;
-    if old_script_exists {
-        write_host_text_file(&new_script_path, &new_script_content, 0o755)?;
-        script_renamed = true;
-    }
-
-    let mut desktop_entry_renamed = false;
-    if old_desktop_exists {
-        write_host_text_file(&new_desktop_entry_path, &new_desktop_content, 0o644)?;
-        desktop_entry_renamed = true;
-    }
-
-    let mut old_script_cleanup_warning = None;
-    let mut old_desktop_entry_cleanup_warning = None;
-
-    // If old paths differ from new paths (slug changed), delete old files
-    let slug_changed = old_launcher_slug != new_slug;
-    if slug_changed {
-        if old_script_exists {
-            old_script_cleanup_warning =
-                remove_old_launcher_file(&old_script_path, SCRIPT_WATERMARK, "script");
-        }
-        if old_desktop_exists {
-            old_desktop_entry_cleanup_warning = remove_old_launcher_file(
-                &old_desktop_entry_path,
-                DESKTOP_ENTRY_WATERMARK,
-                "desktop entry",
-            );
-        }
-    }
-
-    Ok(LauncherRenameResult {
-        old_slug: old_launcher_slug.to_string(),
-        new_slug,
-        new_script_path,
-        new_desktop_entry_path,
-        script_renamed,
-        desktop_entry_renamed,
-        old_script_cleanup_warning,
-        old_desktop_entry_cleanup_warning,
-    })
-}
-
-/// Lists all launchers found in the launchers directory.
-///
-/// Scans `{home}/.local/share/crosshook/launchers/` for files ending in `-trainer.sh`,
-/// derives the slug from each filename, and checks for a matching `.desktop` entry.
-/// Attempts to extract the display name from the `Name=` line in the `.desktop` file.
-/// `is_stale` is reported as `false` for these results because the function does not
-/// have the profile metadata required to compute freshness against an expected name.
-pub fn list_launchers(
-    target_home_path: &str,
-    steam_client_install_path: &str,
-) -> Vec<LauncherInfo> {
-    let home = resolve_target_home_path(target_home_path, steam_client_install_path);
-    let launchers_dir = combine_host_unix_path(&home, ".local/share/crosshook/launchers", "");
-
-    if launchers_dir.is_empty() {
-        return Vec::new();
-    }
-
-    let dir_entries = match fs::read_dir(&launchers_dir) {
-        Ok(entries) => entries,
-        Err(error) if error.kind() == io::ErrorKind::NotFound => return Vec::new(),
-        Err(error) => {
-            tracing::warn!(
-                path = %launchers_dir,
-                %error,
-                "failed to read launcher directory"
-            );
-            return Vec::new();
-        }
-    };
-
-    let mut launchers: Vec<LauncherInfo> = Vec::new();
-
-    for entry in dir_entries {
-        let entry = match entry {
-            Ok(entry) => entry,
-            Err(error) => {
-                tracing::warn!(%error, "failed to read launcher directory entry");
-                continue;
-            }
-        };
-        let file_name = entry.file_name();
-        let file_name_str = file_name.to_string_lossy();
-
-        if !file_name_str.ends_with("-trainer.sh") {
-            continue;
-        }
-
-        // Derive slug by stripping the -trainer.sh suffix
-        let slug = file_name_str
-            .strip_suffix("-trainer.sh")
-            .unwrap_or_default()
-            .to_string();
-
-        if slug.is_empty() {
-            continue;
-        }
-
-        let script_path =
-            combine_host_unix_path(&home, ".local/share/crosshook/launchers", &file_name_str);
-
-        let desktop_entry_path = combine_host_unix_path(
-            &home,
-            ".local/share/applications",
-            &format!("crosshook-{slug}-trainer.desktop"),
-        );
-
-        let script_exists = Path::new(&script_path).exists();
-        let desktop_entry_exists = Path::new(&desktop_entry_path).exists();
-
-        // Try to extract display name from the Name= line in the .desktop file
-        let display_name = if desktop_entry_exists {
-            match extract_display_name_from_desktop(&desktop_entry_path) {
-                Ok(Some(display_name)) => display_name,
-                Ok(None) => slug.clone(),
-                Err(error) => {
-                    tracing::warn!(
-                        path = %desktop_entry_path,
-                        %error,
-                        "failed to inspect launcher desktop entry"
-                    );
-                    slug.clone()
-                }
-            }
-        } else {
-            slug.clone()
-        };
-
-        launchers.push(LauncherInfo {
-            display_name,
-            launcher_slug: slug,
-            script_path,
-            desktop_entry_path,
-            script_exists,
-            desktop_entry_exists,
-            is_stale: false,
-        });
-    }
-
-    launchers.sort_by(|a, b| a.launcher_slug.cmp(&b.launcher_slug));
-    launchers
-}
-
-/// Returns launchers that don't match any known profile slug.
-pub fn find_orphaned_launchers(
-    known_profile_slugs: &[String],
-    target_home_path: &str,
-    steam_client_install_path: &str,
-) -> Vec<LauncherInfo> {
-    let all_launchers = list_launchers(target_home_path, steam_client_install_path);
-    all_launchers
-        .into_iter()
-        .filter(|launcher| !known_profile_slugs.contains(&launcher.launcher_slug))
-        .collect()
-}
-
-fn extract_display_name_from_desktop(desktop_path: &str) -> Result<Option<String>, io::Error> {
-    let content = fs::read_to_string(desktop_path)?;
-    Ok(parse_display_name_from_desktop_content(&content))
-}
-
-fn parse_display_name_from_desktop_content(content: &str) -> Option<String> {
-    for line in content.lines() {
-        if let Some(name_value) = line.strip_prefix("Name=") {
-            return Some(strip_trainer_suffix(name_value));
-        }
-    }
-
-    None
-}
-
-/// Verifies that a file is safe to delete:
-/// 1. Must be a regular file (not a symlink or directory)
-/// 2. Must contain the CrossHook watermark
-///
-/// Returns `Ok(None)` if safe to delete or the file does not exist.
-/// Returns `Ok(Some(reason))` if the file should be skipped.
-fn verify_crosshook_file(path: &str, watermark: &str) -> Result<Option<String>, io::Error> {
-    let metadata = match fs::symlink_metadata(path) {
-        Ok(m) => m,
-        Err(e) if e.kind() == io::ErrorKind::NotFound => return Ok(None),
-        Err(e) => return Err(e),
-    };
-
-    if !metadata.file_type().is_file() {
-        return Ok(Some(format!("Not a regular file: {path}")));
-    }
-
-    let content = fs::read_to_string(path)?;
-    if !content.contains(watermark) {
-        return Ok(Some(format!("Missing CrossHook watermark in: {path}")));
-    }
-
-    Ok(None)
-}
-
-fn remove_old_launcher_file(path: &str, watermark: &str, artifact_label: &str) -> Option<String> {
-    match verify_crosshook_file(path, watermark) {
-        Ok(Some(reason)) => {
-            let warning = format!("Skipped old {artifact_label} cleanup: {reason}");
-            tracing::warn!(
-                path = %path,
-                artifact = artifact_label,
-                reason = %reason,
-                "skipping old launcher cleanup during rename"
-            );
-            Some(warning)
-        }
-        Ok(None) => {
-            if let Err(error) = remove_file_if_exists(path) {
-                let warning = format!("Failed to remove old {artifact_label} at {path}: {error}");
-                tracing::warn!(
-                    path = %path,
-                    artifact = artifact_label,
-                    %error,
-                    "failed to remove old launcher during rename"
-                );
-                Some(warning)
-            } else {
-                None
-            }
-        }
-        Err(error) => {
-            let warning = format!("Failed to verify old {artifact_label} at {path}: {error}");
-            tracing::warn!(
-                path = %path,
-                artifact = artifact_label,
-                %error,
-                "failed to verify old launcher during rename"
-            );
-            Some(warning)
-        }
-    }
-}
-
-/// Attempts to remove a file. Returns `true` if the file was actually removed,
-/// `false` if it did not exist. Treats `ErrorKind::NotFound` as success (idempotent).
-fn remove_file_if_exists(path: &str) -> Result<bool, io::Error> {
-    if path.is_empty() {
-        return Ok(false);
-    }
-
-    match fs::remove_file(path) {
-        Ok(()) => Ok(true),
-        Err(error) if error.kind() == io::ErrorKind::NotFound => Ok(false),
-        Err(error) => Err(error),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::profile::{
-        GameSection, LaunchSection, LauncherSection, SteamSection, TrainerSection,
-    };
-    use std::fs;
-    use tempfile::tempdir;
-
-    /// Helper: use `check_launcher_exists` to derive the canonical paths for a given
-    /// display name and temp home, then return (script_path, desktop_entry_path).
-    fn derive_test_paths(home: &str, display_name: &str) -> (String, String) {
-        let info = check_launcher_exists(display_name, "", "/fake/trainer.exe", home, "")
-            .expect("derive test paths");
-        (info.script_path, info.desktop_entry_path)
-    }
-
-    /// Helper: create parent directories and write a watermarked script file.
-    fn create_watermarked_script(path: &str) {
-        let parent = Path::new(path).parent().expect("parent dir");
-        fs::create_dir_all(parent).expect("create dirs");
-        fs::write(
-            path,
-            "#!/usr/bin/env bash\n# Generated by CrossHook\necho hello\n",
-        )
-        .expect("write file");
-    }
-
-    /// Helper: create parent directories and write a watermarked desktop entry file.
-    fn create_watermarked_desktop(path: &str) {
-        let parent = Path::new(path).parent().expect("parent dir");
-        fs::create_dir_all(parent).expect("create dirs");
-        fs::write(
-            path,
-            "[Desktop Entry]\nName=Test - Trainer\nComment=Trainer launcher for Test. Generated by CrossHook: https://github.com/yandy-r/crosshook\n",
-        )
-        .expect("write file");
-    }
-
-    /// Helper: create parent directories and write a placeholder file (no watermark).
-    fn create_file_at(path: &str) {
-        let parent = Path::new(path).parent().expect("parent dir");
-        fs::create_dir_all(parent).expect("create dirs");
-        fs::write(path, "placeholder").expect("write file");
-    }
-
-    /// Helper: create parent directories and write a file with specific content.
-    fn create_file_with_content(path: &str, content: &str) {
-        let parent = Path::new(path).parent().expect("parent dir");
-        fs::create_dir_all(parent).expect("create dirs");
-        fs::write(path, content).expect("write file");
-    }
-
-    /// Helper: construct a minimal `SteamExternalLauncherExportRequest` for rename tests.
-    fn make_test_request() -> SteamExternalLauncherExportRequest {
-        SteamExternalLauncherExportRequest {
-            method: "proton_run".to_string(),
-            launcher_name: String::new(),
-            trainer_path: "/opt/trainers/TestTrainer.exe".to_string(),
-            trainer_loading_mode: crate::profile::TrainerLoadingMode::SourceDirectory,
-            launcher_icon_path: String::new(),
-            prefix_path: "/tmp/prefix".to_string(),
-            proton_path: "/opt/proton/proton".to_string(),
-            steam_app_id: String::new(),
-            steam_client_install_path: String::new(),
-            target_home_path: String::new(),
-            profile_name: None,
-            ..Default::default()
-        }
-    }
-
-    // --- check_launcher_exists tests ---
-
-    #[test]
-    fn check_when_both_files_exist() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let (script_path, desktop_path) = derive_test_paths(&home, "Test Game");
-
-        create_file_at(&script_path);
-        create_file_at(&desktop_path);
-
-        let info = check_launcher_exists("Test Game", "", "/fake/trainer.exe", &home, "")
-            .expect("check launcher exists");
-        assert!(info.script_exists, "script should exist");
-        assert!(info.desktop_entry_exists, "desktop entry should exist");
-        assert!(!info.script_path.is_empty());
-        assert!(!info.desktop_entry_path.is_empty());
-        assert!(
-            info.is_stale,
-            "placeholder desktop content should be treated as stale"
-        );
-    }
-
-    #[test]
-    fn check_when_neither_exists() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-
-        let info = check_launcher_exists("Test Game", "", "/fake/trainer.exe", &home, "")
-            .expect("check launcher exists");
-        assert!(!info.script_exists, "script should not exist");
-        assert!(!info.desktop_entry_exists, "desktop entry should not exist");
-    }
-
-    #[test]
-    fn check_launcher_exists_for_request_marks_mode_mismatches_as_stale() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let display_name = "Aurora Test";
-        let request = SteamExternalLauncherExportRequest {
-            method: "proton_run".to_string(),
-            launcher_name: display_name.to_string(),
-            trainer_path: "/opt/trainers/Aurora.exe".to_string(),
-            trainer_loading_mode: crate::profile::TrainerLoadingMode::SourceDirectory,
-            launcher_icon_path: String::new(),
-            prefix_path: "/games/prefixes/aurora-test".to_string(),
-            proton_path: "/opt/proton/proton".to_string(),
-            steam_app_id: String::new(),
-            steam_client_install_path: String::new(),
-            target_home_path: home.clone(),
-            profile_name: None,
-            ..Default::default()
-        };
-
-        let info = check_launcher_exists_for_request(display_name, &request)
-            .expect("derive launcher paths");
-        let copy_request = SteamExternalLauncherExportRequest {
-            trainer_loading_mode: crate::profile::TrainerLoadingMode::CopyToPrefix,
-            ..request.clone()
-        };
-
-        create_file_with_content(
-            &info.script_path,
-            &build_trainer_script_content(&copy_request, display_name),
-        );
-        create_file_with_content(
-            &info.desktop_entry_path,
-            &build_desktop_entry_content(display_name, &info.launcher_slug, &info.script_path, ""),
-        );
-
-        let info = check_launcher_exists_for_request(display_name, &request)
-            .expect("check launcher request");
-        assert!(info.script_exists);
-        assert!(info.desktop_entry_exists);
-        assert!(info.is_stale, "mode mismatch should mark launcher stale");
-    }
-
-    #[test]
-    fn check_when_only_script_exists() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let (script_path, _desktop_path) = derive_test_paths(&home, "Partial Game");
-
-        create_file_at(&script_path);
-
-        let info = check_launcher_exists("Partial Game", "", "/fake/trainer.exe", &home, "")
-            .expect("check launcher exists");
-        assert!(info.script_exists, "script should exist");
-        assert!(!info.desktop_entry_exists, "desktop entry should not exist");
-    }
-
-    #[test]
-    fn check_marks_launcher_stale_when_display_name_mismatches() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let (script_path, desktop_path) = derive_test_paths(&home, "Expected Game");
-
-        create_file_at(&script_path);
-        create_file_with_content(
-            &desktop_path,
-            "[Desktop Entry]\nName=Different Game - Trainer\nExec=/bin/bash trainer.sh\n",
-        );
-
-        let info = check_launcher_exists("Expected Game", "", "/fake/trainer.exe", &home, "")
-            .expect("check launcher exists");
-
-        assert!(info.is_stale);
-    }
-
-    #[test]
-    fn check_returns_error_when_desktop_entry_cannot_be_read() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let (script_path, desktop_path) = derive_test_paths(&home, "Encoded Game");
-
-        create_file_at(&script_path);
-        let parent = Path::new(&desktop_path).parent().expect("desktop parent");
-        fs::create_dir_all(parent).expect("desktop parent dirs");
-        fs::write(&desktop_path, [0xff, 0xfe, 0xfd]).expect("write invalid desktop bytes");
-
-        let error = check_launcher_exists("Encoded Game", "", "/fake/trainer.exe", &home, "")
-            .expect_err("invalid desktop bytes should now surface as an error");
-
-        assert!(matches!(error, LauncherStoreError::Io(_)));
-    }
-
-    // --- delete_launcher_files tests ---
-
-    #[test]
-    fn delete_when_both_exist() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let (script_path, desktop_path) = derive_test_paths(&home, "Delete Both");
-
-        create_watermarked_script(&script_path);
-        create_watermarked_desktop(&desktop_path);
-
-        let result = delete_launcher_files("Delete Both", "", "/fake/trainer.exe", &home, "")
-            .expect("delete should succeed");
-
-        assert!(result.script_deleted, "script should be deleted");
-        assert!(
-            result.desktop_entry_deleted,
-            "desktop entry should be deleted"
-        );
-        assert!(
-            !Path::new(&result.script_path).exists(),
-            "script file should be gone"
-        );
-        assert!(
-            !Path::new(&result.desktop_entry_path).exists(),
-            "desktop entry file should be gone"
-        );
-    }
-
-    #[test]
-    fn delete_when_neither_exists_is_noop() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-
-        let result = delete_launcher_files("No Files", "", "/fake/trainer.exe", &home, "")
-            .expect("delete should succeed even with no files");
-
-        assert!(!result.script_deleted, "nothing to delete for script");
-        assert!(
-            !result.desktop_entry_deleted,
-            "nothing to delete for desktop entry"
-        );
-    }
-
-    #[test]
-    fn delete_when_only_script_exists() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let (script_path, _desktop_path) = derive_test_paths(&home, "Script Only");
-
-        create_watermarked_script(&script_path);
-
-        let result = delete_launcher_files("Script Only", "", "/fake/trainer.exe", &home, "")
-            .expect("delete should succeed");
-
-        assert!(result.script_deleted, "script should be deleted");
-        assert!(
-            !result.desktop_entry_deleted,
-            "desktop entry was not present"
-        );
-        assert!(
-            !Path::new(&result.script_path).exists(),
-            "script file should be gone"
-        );
-    }
-
-    #[test]
-    fn delete_launcher_for_profile_delegates_correctly() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-
-        let profile = GameProfile {
-            game: GameSection {
-                name: "Test Game".to_string(),
-                executable_path: String::new(),
-                custom_cover_art_path: String::new(),
-                custom_portrait_art_path: String::new(),
-                custom_background_art_path: String::new(),
-            },
-            trainer: TrainerSection {
-                path: "/mnt/trainers/test.exe".to_string(),
-                kind: String::new(),
-                loading_mode: crate::profile::TrainerLoadingMode::SourceDirectory,
-                trainer_type: "unknown".to_string(),
-                required_protontricks: Vec::new(),
-                community_trainer_sha256: String::new(),
-            },
-            steam: SteamSection {
-                app_id: "12345".to_string(),
-                launcher: LauncherSection {
-                    display_name: "Test Game".to_string(),
-                    icon_path: String::new(),
-                },
-                ..Default::default()
-            },
-            launch: LaunchSection {
-                method: "steam_applaunch".to_string(),
-                ..Default::default()
-            },
-            ..Default::default()
-        };
-
-        // Derive paths using the same inputs the profile facade will use
-        let info = check_launcher_exists(
-            &profile.steam.launcher.display_name,
-            &profile.steam.app_id,
-            &profile.trainer.path,
-            &home,
-            "",
-        )
-        .expect("derive launcher paths");
-
-        create_watermarked_script(&info.script_path);
-        create_watermarked_desktop(&info.desktop_entry_path);
-
-        let result = delete_launcher_for_profile(&profile, &home, "")
-            .expect("delete for profile should succeed");
-
-        assert!(result.script_deleted);
-        assert!(result.desktop_entry_deleted);
-        assert!(!Path::new(&result.script_path).exists());
-        assert!(!Path::new(&result.desktop_entry_path).exists());
-    }
-
-    #[test]
-    fn check_launcher_for_profile_delegates_correctly() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-
-        let profile = GameProfile {
-            game: GameSection {
-                name: "Test Game".to_string(),
-                executable_path: String::new(),
-                custom_cover_art_path: String::new(),
-                custom_portrait_art_path: String::new(),
-                custom_background_art_path: String::new(),
-            },
-            trainer: TrainerSection {
-                path: "/mnt/trainers/test.exe".to_string(),
-                kind: String::new(),
-                loading_mode: crate::profile::TrainerLoadingMode::SourceDirectory,
-                trainer_type: "unknown".to_string(),
-                required_protontricks: Vec::new(),
-                community_trainer_sha256: String::new(),
-            },
-            steam: SteamSection {
-                app_id: "12345".to_string(),
-                launcher: LauncherSection {
-                    display_name: "Test Game".to_string(),
-                    icon_path: String::new(),
-                },
-                ..Default::default()
-            },
-            launch: LaunchSection {
-                method: "steam_applaunch".to_string(),
-                ..Default::default()
-            },
-            ..Default::default()
-        };
-
-        let direct = check_launcher_exists(
-            &profile.steam.launcher.display_name,
-            &profile.steam.app_id,
-            &profile.trainer.path,
-            &home,
-            "",
-        )
-        .expect("check launcher exists");
-        let delegated = check_launcher_for_profile(&profile, &home, "", UmuPreference::Auto)
-            .expect("check launcher for profile");
-
-        assert_eq!(delegated, direct);
-    }
-
-    #[test]
-    fn check_launcher_for_profile_resolves_legacy_empty_method_as_steam() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-
-        let profile = GameProfile {
-            game: GameSection {
-                name: "Legacy Steam Game".to_string(),
-                executable_path: "/games/legacy/launcher.exe".to_string(),
-                custom_cover_art_path: String::new(),
-                custom_portrait_art_path: String::new(),
-                custom_background_art_path: String::new(),
-            },
-            trainer: TrainerSection {
-                path: "/mnt/trainers/legacy.exe".to_string(),
-                kind: String::new(),
-                loading_mode: crate::profile::TrainerLoadingMode::SourceDirectory,
-                trainer_type: "unknown".to_string(),
-                required_protontricks: Vec::new(),
-                community_trainer_sha256: String::new(),
-            },
-            steam: SteamSection {
-                enabled: true,
-                app_id: "12345".to_string(),
-                compatdata_path: "/steam/compatdata/12345".to_string(),
-                proton_path: "/steam/proton/proton".to_string(),
-                launcher: LauncherSection {
-                    display_name: "Legacy Steam Game".to_string(),
-                    icon_path: String::new(),
-                },
-            },
-            runtime: crate::profile::RuntimeSection {
-                prefix_path: "/wrong/runtime/prefix".to_string(),
-                proton_path: "/wrong/runtime/proton".to_string(),
-                working_directory: String::new(),
-                steam_app_id: String::new(),
-                umu_game_id: String::new(),
-                umu_preference: None,
-            },
-            launch: LaunchSection {
-                method: String::new(),
-                ..Default::default()
-            },
-            ..Default::default()
-        };
-
-        let direct = check_launcher_exists_for_request(
-            &profile.steam.launcher.display_name,
-            &SteamExternalLauncherExportRequest {
-                method: "steam_applaunch".to_string(),
-                launcher_name: profile.steam.launcher.display_name.clone(),
-                trainer_path: profile.trainer.path.clone(),
-                trainer_loading_mode: profile.trainer.loading_mode,
-                launcher_icon_path: profile.steam.launcher.icon_path.clone(),
-                prefix_path: profile.steam.compatdata_path.clone(),
-                proton_path: profile.steam.proton_path.clone(),
-                steam_app_id: profile.steam.app_id.clone(),
-                steam_client_install_path: String::new(),
-                target_home_path: home.clone(),
-                profile_name: None,
-                ..Default::default()
-            },
-        )
-        .expect("check launcher exists for explicit steam request");
-
-        let delegated = check_launcher_for_profile(&profile, &home, "", UmuPreference::Auto)
-            .expect("check launcher for profile");
-
-        assert_eq!(delegated, direct);
-    }
-
-    #[test]
-    fn delete_launcher_by_slug_deletes_matching_files() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-
-        let (script_path, desktop_path) =
-            derive_launcher_paths_from_slug("delete-by-slug", &home, "");
-
-        create_watermarked_script(&script_path);
-        create_watermarked_desktop(&desktop_path);
-
-        let result = delete_launcher_by_slug("delete-by-slug", &home, "")
-            .expect("delete by slug should succeed");
-
-        assert!(result.script_deleted);
-        assert!(result.desktop_entry_deleted);
-        assert_eq!(result.script_path, script_path);
-        assert_eq!(result.desktop_entry_path, desktop_path);
-        assert!(!Path::new(&result.script_path).exists());
-        assert!(!Path::new(&result.desktop_entry_path).exists());
-    }
-
-    // --- rename_launcher_files tests ---
-
-    #[test]
-    fn rename_when_old_exists() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let request = make_test_request();
-
-        let old_slug = "old-game";
-        let old_script = combine_host_unix_path(
-            &home,
-            ".local/share/crosshook/launchers",
-            &format!("{old_slug}-trainer.sh"),
-        );
-        let old_desktop = combine_host_unix_path(
-            &home,
-            ".local/share/applications",
-            &format!("crosshook-{old_slug}-trainer.desktop"),
-        );
-
-        create_watermarked_script(&old_script);
-        create_watermarked_desktop(&old_desktop);
-
-        let result = rename_launcher_files(old_slug, "New Game", "", &home, "", &request)
-            .expect("rename should succeed");
-
-        assert_eq!(result.old_slug, "old-game");
-        assert_eq!(result.new_slug, "new-game");
-        assert!(result.script_renamed, "script should be renamed");
-        assert!(
-            result.desktop_entry_renamed,
-            "desktop entry should be renamed"
-        );
-
-        // Old files should be deleted (slug changed)
-        assert!(
-            !Path::new(&old_script).exists(),
-            "old script should be removed"
-        );
-        assert!(
-            !Path::new(&old_desktop).exists(),
-            "old desktop entry should be removed"
-        );
-
-        // New files should exist with updated content
-        assert!(
-            Path::new(&result.new_script_path).exists(),
-            "new script should exist"
-        );
-        assert!(
-            Path::new(&result.new_desktop_entry_path).exists(),
-            "new desktop entry should exist"
-        );
-
-        let new_script_content =
-            fs::read_to_string(&result.new_script_path).expect("read new script");
-        assert!(
-            new_script_content.contains("# New Game - Trainer launcher"),
-            "new script should contain updated display name"
-        );
-
-        let new_desktop_content =
-            fs::read_to_string(&result.new_desktop_entry_path).expect("read new desktop");
-        assert!(
-            new_desktop_content.contains("Name=New Game - Trainer"),
-            "new desktop should contain updated display name"
-        );
-    }
-
-    #[test]
-    fn rename_when_old_does_not_exist() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let request = make_test_request();
-
-        let result = rename_launcher_files("nonexistent-game", "New Name", "", &home, "", &request)
-            .expect("rename should succeed with no-op");
-
-        assert!(!result.script_renamed, "script should not be renamed");
-        assert!(
-            !result.desktop_entry_renamed,
-            "desktop entry should not be renamed"
-        );
-        assert_eq!(result.old_slug, "nonexistent-game");
-        assert_eq!(result.new_slug, "new-name");
-    }
-
-    #[test]
-    fn rename_when_slug_unchanged() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let request = make_test_request();
-
-        // The slug for "My Game" is "my-game". If we rename from "my-game" to "My Game",
-        // the slug stays the same -- content should be rewritten in place without deleting.
-        let slug = "my-game";
-        let script_path = combine_host_unix_path(
-            &home,
-            ".local/share/crosshook/launchers",
-            &format!("{slug}-trainer.sh"),
-        );
-        let desktop_path = combine_host_unix_path(
-            &home,
-            ".local/share/applications",
-            &format!("crosshook-{slug}-trainer.desktop"),
-        );
-
-        create_file_with_content(&script_path, "#!/usr/bin/env bash\n# old content");
-        create_file_with_content(
-            &desktop_path,
-            "[Desktop Entry]\nName=Old Name - Trainer\nExec=/bin/bash old.sh\n",
-        );
-
-        let result = rename_launcher_files(slug, "My Game", "", &home, "", &request)
-            .expect("rename should succeed");
-
-        assert_eq!(result.old_slug, "my-game");
-        assert_eq!(result.new_slug, "my-game");
-        assert!(result.script_renamed, "script should be rewritten");
-        assert!(
-            result.desktop_entry_renamed,
-            "desktop entry should be rewritten"
-        );
-
-        // Files should still exist at the same paths (not deleted)
-        assert!(
-            Path::new(&script_path).exists(),
-            "script should still exist at same path"
-        );
-        assert!(
-            Path::new(&desktop_path).exists(),
-            "desktop entry should still exist at same path"
-        );
-
-        // Content should be updated
-        let new_script_content = fs::read_to_string(&script_path).expect("read script");
-        assert!(
-            new_script_content.contains("# My Game - Trainer launcher"),
-            "script content should be updated with new display name"
-        );
-
-        let new_desktop_content = fs::read_to_string(&desktop_path).expect("read desktop");
-        assert!(
-            new_desktop_content.contains("Name=My Game - Trainer"),
-            "desktop content should be updated with new display name"
-        );
-    }
-
-    #[test]
-    fn rename_when_only_script_exists() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let request = make_test_request();
-        let old_slug = "script-only";
-        let old_script = combine_host_unix_path(
-            &home,
-            ".local/share/crosshook/launchers",
-            &format!("{old_slug}-trainer.sh"),
-        );
-
-        create_watermarked_script(&old_script);
-
-        let result =
-            rename_launcher_files(old_slug, "Script Only Renamed", "", &home, "", &request)
-                .expect("rename should succeed");
-
-        assert!(result.script_renamed);
-        assert!(!result.desktop_entry_renamed);
-        assert!(Path::new(&result.new_script_path).exists());
-        assert!(!Path::new(&old_script).exists());
-    }
-
-    #[test]
-    fn rename_when_only_desktop_entry_exists() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let request = make_test_request();
-        let old_slug = "desktop-only";
-        let old_desktop = combine_host_unix_path(
-            &home,
-            ".local/share/applications",
-            &format!("crosshook-{old_slug}-trainer.desktop"),
-        );
-
-        create_watermarked_desktop(&old_desktop);
-
-        let result =
-            rename_launcher_files(old_slug, "Desktop Only Renamed", "", &home, "", &request)
-                .expect("rename should succeed");
-
-        assert!(!result.script_renamed);
-        assert!(result.desktop_entry_renamed);
-        assert!(Path::new(&result.new_desktop_entry_path).exists());
-        assert!(!Path::new(&old_desktop).exists());
-    }
-
-    #[test]
-    fn rename_reports_warning_when_old_file_is_not_crosshook_managed() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let request = make_test_request();
-        let old_slug = "unsafe-old";
-        let old_script = combine_host_unix_path(
-            &home,
-            ".local/share/crosshook/launchers",
-            &format!("{old_slug}-trainer.sh"),
-        );
-        let old_desktop = combine_host_unix_path(
-            &home,
-            ".local/share/applications",
-            &format!("crosshook-{old_slug}-trainer.desktop"),
-        );
-
-        create_file_with_content(&old_script, "#!/usr/bin/env bash\necho unmanaged\n");
-        create_watermarked_desktop(&old_desktop);
-
-        let result = rename_launcher_files(old_slug, "Safe New Name", "", &home, "", &request)
-            .expect("rename should succeed");
-
-        assert!(result.script_renamed);
-        assert!(result.desktop_entry_renamed);
-        assert!(result.old_script_cleanup_warning.is_some());
-        assert!(result.old_desktop_entry_cleanup_warning.is_none());
-        assert!(
-            Path::new(&old_script).exists(),
-            "unsafe old script should remain"
-        );
-    }
-
-    // --- list_launchers tests ---
-
-    #[test]
-    fn list_with_no_launchers() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-
-        let result = list_launchers(&home, "");
-        assert!(result.is_empty(), "should return empty vec");
-    }
-
-    #[test]
-    fn list_with_launchers() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-
-        // Create two launcher scripts
-        let script_a = combine_host_unix_path(
-            &home,
-            ".local/share/crosshook/launchers",
-            "alpha-game-trainer.sh",
-        );
-        let desktop_a = combine_host_unix_path(
-            &home,
-            ".local/share/applications",
-            "crosshook-alpha-game-trainer.desktop",
-        );
-        let script_b = combine_host_unix_path(
-            &home,
-            ".local/share/crosshook/launchers",
-            "beta-game-trainer.sh",
-        );
-
-        create_file_with_content(&script_a, "#!/usr/bin/env bash\n# script a");
-        create_file_with_content(
-            &desktop_a,
-            "[Desktop Entry]\nName=Alpha Game - Trainer\nExec=/bin/bash alpha.sh\n",
-        );
-        create_file_with_content(&script_b, "#!/usr/bin/env bash\n# script b");
-
-        let result = list_launchers(&home, "");
-
-        assert_eq!(result.len(), 2, "should find two launchers");
-
-        // Results should be sorted by slug
-        assert_eq!(result[0].launcher_slug, "alpha-game");
-        assert_eq!(result[0].display_name, "Alpha Game");
-        assert!(result[0].script_exists);
-        assert!(result[0].desktop_entry_exists);
-
-        assert_eq!(result[1].launcher_slug, "beta-game");
-        // No .desktop file, so display_name falls back to slug
-        assert_eq!(result[1].display_name, "beta-game");
-        assert!(result[1].script_exists);
-        assert!(!result[1].desktop_entry_exists);
-    }
-
-    #[test]
-    fn find_orphaned_launchers_returns_only_unknown_slugs() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-
-        create_file_with_content(
-            &combine_host_unix_path(
-                &home,
-                ".local/share/crosshook/launchers",
-                "known-game-trainer.sh",
-            ),
-            "#!/usr/bin/env bash\n# script known",
-        );
-        create_file_with_content(
-            &combine_host_unix_path(
-                &home,
-                ".local/share/crosshook/launchers",
-                "orphan-game-trainer.sh",
-            ),
-            "#!/usr/bin/env bash\n# script orphan",
-        );
-
-        let result = find_orphaned_launchers(&["known-game".to_string()], &home, "");
-
-        assert_eq!(result.len(), 1);
-        assert_eq!(result[0].launcher_slug, "orphan-game");
-    }
-
-    #[test]
-    fn parse_display_name_handles_name_without_suffix() {
-        assert_eq!(
-            parse_display_name_from_desktop_content("[Desktop Entry]\nName=Standalone Launcher\n"),
-            Some("Standalone Launcher".to_string())
-        );
-    }
-
-    #[test]
-    fn parse_display_name_returns_none_without_name_line() {
-        assert_eq!(
-            parse_display_name_from_desktop_content(
-                "[Desktop Entry]\nExec=/bin/bash launcher.sh\n"
-            ),
-            None
-        );
-    }
-
-    // --- watermark verification tests ---
-
-    #[test]
-    fn delete_skips_symlink() {
-        use std::os::unix::fs as unix_fs;
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let (script_path, _desktop_path) = derive_test_paths(&home, "Symlink Test");
-
-        let real_file = temp.path().join("real-script.sh");
-        fs::write(
-            &real_file,
-            "#!/usr/bin/env bash\n# Generated by CrossHook\n",
-        )
-        .expect("write real");
-        let parent = Path::new(&script_path).parent().expect("parent");
-        fs::create_dir_all(parent).expect("mkdir");
-        unix_fs::symlink(&real_file, &script_path).expect("symlink");
-
-        let result = delete_launcher_files("Symlink Test", "", "/fake/trainer.exe", &home, "")
-            .expect("delete should succeed");
-
-        assert!(
-            !result.script_deleted,
-            "symlinked script should not be deleted"
-        );
-        assert!(
-            result.script_skipped_reason.is_some(),
-            "should have skip reason for symlink"
-        );
-        assert!(result
-            .script_skipped_reason
-            .as_ref()
-            .unwrap()
-            .contains("Not a regular file"),);
-    }
-
-    #[test]
-    fn delete_skips_file_without_watermark() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let (script_path, desktop_path) = derive_test_paths(&home, "No Watermark");
-
-        create_file_at(&script_path); // plain "placeholder", no watermark
-        create_file_at(&desktop_path);
-
-        let result = delete_launcher_files("No Watermark", "", "/fake/trainer.exe", &home, "")
-            .expect("delete should succeed");
-
-        assert!(!result.script_deleted);
-        assert!(!result.desktop_entry_deleted);
-        assert!(result.script_skipped_reason.is_some());
-        assert!(result.desktop_entry_skipped_reason.is_some());
-        assert!(Path::new(&script_path).exists(), "file should still exist");
-        assert!(Path::new(&desktop_path).exists(), "file should still exist");
-    }
-
-    #[test]
-    fn delete_proceeds_with_watermarked_file() {
-        let temp = tempdir().expect("temp dir");
-        let home = temp.path().to_string_lossy().into_owned();
-        let (script_path, desktop_path) = derive_test_paths(&home, "Watermarked");
-
-        create_watermarked_script(&script_path);
-        create_watermarked_desktop(&desktop_path);
-
-        let result = delete_launcher_files("Watermarked", "", "/fake/trainer.exe", &home, "")
-            .expect("delete should succeed");
-
-        assert!(result.script_deleted);
-        assert!(result.desktop_entry_deleted);
-        assert!(result.script_skipped_reason.is_none());
-        assert!(result.desktop_entry_skipped_reason.is_none());
-        assert!(!Path::new(&script_path).exists());
-        assert!(!Path::new(&desktop_path).exists());
-    }
-}
