+++ src/crosshook-native/crates/crosshook-core/src/export/launcher_store/mod.rs
@@ -1,0 +1,32 @@
+//! Launcher store — per-domain submodules for the launcher file lifecycle.
+//!
+//! - `types`     — public DTOs and error type
+//! - `paths`     — path derivation helpers
+//! - `fs_ops`    — low-level filesystem primitives (watermark, remove, name parsing)
+//! - `queries`   — read-only launcher existence checks, listing, orphan detection
+//! - `mutations` — delete and rename operations
+
+mod fs_ops;
+mod mutations;
+mod paths;
+mod queries;
+mod types;
+
+#[cfg(test)]
+mod tests;
+
+/// Watermark present in generated `.sh` trainer launcher scripts.
+pub(crate) const SCRIPT_WATERMARK: &str = "# Generated by CrossHook";
+
+/// Watermark present in generated `.desktop` entry files.
+pub(crate) const DESKTOP_ENTRY_WATERMARK: &str = "Generated by CrossHook";
+
+pub use mutations::{
+    delete_launcher_by_slug, delete_launcher_files, delete_launcher_for_profile,
+    rename_launcher_files,
+};
+pub use queries::{
+    check_launcher_exists, check_launcher_exists_for_request, check_launcher_for_profile,
+    find_orphaned_launchers, list_launchers,
+};
+pub use types::{LauncherDeleteResult, LauncherInfo, LauncherRenameResult, LauncherStoreError};
