+++ src/crosshook-native/crates/crosshook-core/src/export/launcher_store/tests/mod.rs
@@ -1,0 +1,74 @@
+mod check;
+mod delete;
+mod list;
+mod rename;
+
+use std::fs;
+use std::path::Path;
+
+use crate::export::launcher::SteamExternalLauncherExportRequest;
+use crate::profile::TrainerLoadingMode;
+
+use super::queries::check_launcher_exists;
+
+/// Helper: use `check_launcher_exists` to derive the canonical paths for a given
+/// display name and temp home, then return (script_path, desktop_entry_path).
+pub(super) fn derive_test_paths(home: &str, display_name: &str) -> (String, String) {
+    let info = check_launcher_exists(display_name, "", "/fake/trainer.exe", home, "")
+        .expect("derive test paths");
+    (info.script_path, info.desktop_entry_path)
+}
+
+/// Helper: create parent directories and write a watermarked script file.
+pub(super) fn create_watermarked_script(path: &str) {
+    let parent = Path::new(path).parent().expect("parent dir");
+    fs::create_dir_all(parent).expect("create dirs");
+    fs::write(
+        path,
+        "#!/usr/bin/env bash\n# Generated by CrossHook\necho hello\n",
+    )
+    .expect("write file");
+}
+
+/// Helper: create parent directories and write a watermarked desktop entry file.
+pub(super) fn create_watermarked_desktop(path: &str) {
+    let parent = Path::new(path).parent().expect("parent dir");
+    fs::create_dir_all(parent).expect("create dirs");
+    fs::write(
+        path,
+        "[Desktop Entry]\nName=Test - Trainer\nComment=Trainer launcher for Test. Generated by CrossHook: https://github.com/yandy-r/crosshook\n",
+    )
+    .expect("write file");
+}
+
+/// Helper: create parent directories and write a placeholder file (no watermark).
+pub(super) fn create_file_at(path: &str) {
+    let parent = Path::new(path).parent().expect("parent dir");
+    fs::create_dir_all(parent).expect("create dirs");
+    fs::write(path, "placeholder").expect("write file");
+}
+
+/// Helper: create parent directories and write a file with specific content.
+pub(super) fn create_file_with_content(path: &str, content: &str) {
+    let parent = Path::new(path).parent().expect("parent dir");
+    fs::create_dir_all(parent).expect("create dirs");
+    fs::write(path, content).expect("write file");
+}
+
+/// Helper: construct a minimal `SteamExternalLauncherExportRequest` for rename tests.
+pub(super) fn make_test_request() -> SteamExternalLauncherExportRequest {
+    SteamExternalLauncherExportRequest {
+        method: "proton_run".to_string(),
+        launcher_name: String::new(),
+        trainer_path: "/opt/trainers/TestTrainer.exe".to_string(),
+        trainer_loading_mode: TrainerLoadingMode::SourceDirectory,
+        launcher_icon_path: String::new(),
+        prefix_path: "/tmp/prefix".to_string(),
+        proton_path: "/opt/proton/proton".to_string(),
+        steam_app_id: String::new(),
+        steam_client_install_path: String::new(),
+        target_home_path: String::new(),
+        profile_name: None,
+        ..Default::default()
+    }
+}
